pyproject.toml
[project]
name = "text2sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.14.1",
    "ipykernel>=7.1.0",
    "litellm>=1.81.5",
    "matplotlib>=3.10.8",
    "mlflow>=2.10.0",
    "numpy>=2.4.1",
    "pandas<3",
    "protobuf<5",
    "setuptools>=80.10.2",
]
